# backend/rfc-service/src/main/resources/static/openapi.yaml
openapi: 3.0.3
info:
  description: API для системы управления запросами на изменение (RFC)
  title: RFC Management System API
  version: 1.0.0
servers:
- url: /
security:
- bearerAuth: []
paths:
...
  /system:
    get:
      description: Возвращает список систем с поддержкой пагинации и фильтрации по
        названию
      operationId: getSystems
      parameters:
      - description: Номер страницы (начиная с 0)
        in: query
        name: page
        required: false
        schema:
          default: 0
          minimum: 0
          type: integer
      - description: Размер страницы
        in: query
        name: size
        required: false
        schema:
          default: 20
          maximum: 100
          minimum: 1
          type: integer
      - description: Фильтр по названию системы (поиск по совпадению)
        in: query
        name: name
        required: false
        schema:
          minLength: 1
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemPageResponse"
          description: Успешный возврат списка систем
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal Server Error
      summary: Получить пагинированный список систем
      tags:
      - Systems
    post:
      description: Создание новой системы в системе
      operationId: createSystem
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SystemRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemResponse"
          description: Система успешно создана
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unauthorized
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal Server Error
      summary: Создать новую систему
      tags:
      - Systems
  /system/{id}:
    delete:
      description: Удаление системы по идентификатору
      operationId: deleteSystem
      parameters:
      - description: ID элемента
        in: path
        name: id
        required: true
        schema:
          format: int64
          nullable: false
          type: integer
      responses:
        "204":
          description: Система успешно удалена
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal Server Error
      summary: Удалить систему
      tags:
      - Systems
    get:
      description: Возвращает детальную информацию о системе по её идентификатору
      operationId: getSystemById
      parameters:
      - description: ID элемента
        in: path
        name: id
        required: true
        schema:
          format: int64
          nullable: false
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemResponse"
          description: Успешный возврат системы
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal Server Error
      summary: Получить систему по ID
      tags:
      - Systems
    put:
      description: Полное обновление информации о системе
      operationId: updateSystem
      parameters:
      - description: ID элемента
        in: path
        name: id
        required: true
        schema:
          format: int64
          nullable: false
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SystemRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SystemResponse"
          description: Система успешно обновлена
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal Server Error
      summary: Обновить систему
      tags:
      - Systems
  /system/{id}/subsystem:
    get:
      description: Возвращает список подсистем для указанной системы
      operationId: getSystemSubsystems
      parameters:
      - description: ID элемента
        in: path
        name: id
        required: true
        schema:
          format: int64
          nullable: false
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubsystemPageResponse"
          description: Успешный возврат списка подсистем
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal Server Error
      summary: Получить список подсистем системы
      tags:
      - Subsystems
    post:
      description: Создание новой подсистемы для указанной системы
      operationId: createSubsystem
      parameters:
      - description: ID элемента
        in: path
        name: id
        required: true
        schema:
          format: int64
          nullable: false
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubsystemRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubsystemResponse"
          description: Подсистема успешно создана
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal Server Error
      summary: Создать подсистему
      tags:
      - Subsystems
  /system/{id}/subsystem/{subsystemId}:
    delete:
      description: Удаление подсистемы
      operationId: deleteSubsystem
      parameters:
      - description: ID элемента
        in: path
        name: id
        required: true
        schema:
          format: int64
          nullable: false
          type: integer
      - description: ID подсистемы
        in: path
        name: subsystemId
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "204":
          description: Подсистема успешно удалена
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal Server Error
      summary: Удалить подсистему
      tags:
      - Subsystems
    get:
      description: Возвращает детальную информацию о подсистеме
      operationId: getSubsystemById
      parameters:
      - description: ID элемента
        in: path
        name: id
        required: true
        schema:
          format: int64
          nullable: false
          type: integer
      - description: ID подсистемы
        in: path
        name: subsystemId
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubsystemResponse"
          description: Успешный возврат подсистемы
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal Server Error
      summary: Получить подсистему по ID
      tags:
      - Subsystems
    put:
      description: Полное обновление информации о подсистеме
      operationId: updateSubsystem
      parameters:
      - description: ID элемента
        in: path
        name: id
        required: true
        schema:
          format: int64
          nullable: false
          type: integer
      - description: ID подсистемы
        in: path
        name: subsystemId
        required: true
        schema:
          format: int64
          type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubsystemRequest"
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SubsystemResponse"
          description: Подсистема успешно обновлена
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Bad Request
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Unauthorized
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Not Found
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal Server Error
      summary: Обновить подсистему
      tags:
      - Subsystems
...
components:
...
  schemas:
...
    ErrorResponse:
      example:
        errors:
        - code: code
          message: message
        - code: code
          message: message
      nullable: false
      properties:
        errors:
          items:
            $ref: "#/components/schemas/Error"
          type: array
      required:
      - errors
      type: object
    Error:
      example:
        code: code
        message: message
      properties:
        code:
          type: string
        message:
          type: string
      required:
      - code
      - message
      type: object
...
    SystemPageResponse:
      allOf:
      - $ref: "#/components/schemas/PageResponse"
      - properties:
          content:
            items:
              $ref: "#/components/schemas/SystemResponse"
            type: array
        required:
        - content
        type: object
      example:
        number: 5
        size: 1
        last: true
        totalPages: 6
        first: true
        content:
        - name: name
          description: description
          id: 5
          subsystems:
          - systemId: 7
            teamId: 9
            name: name
            description: description
            id: 2
          - systemId: 7
            teamId: 9
            name: name
            description: description
            id: 2
        - name: name
          description: description
          id: 5
          subsystems:
          - systemId: 7
            teamId: 9
            name: name
            description: description
            id: 2
          - systemId: 7
            teamId: 9
            name: name
            description: description
            id: 2
        totalElements: 0
    SystemResponse:
      description: Данные системы
      example:
        name: name
        description: description
        id: 5
        subsystems:
        - systemId: 7
          teamId: 9
          name: name
          description: description
          id: 2
        - systemId: 7
          teamId: 9
          name: name
          description: description
          id: 2
      properties:
        id:
          description: Уникальный идентификатор системы
          format: int64
          type: integer
        name:
          description: Название системы
          type: string
        description:
          description: Описание системы
          nullable: true
          type: string
        subsystems:
          description: Список подсистем
          items:
            $ref: "#/components/schemas/SubsystemResponse"
          type: array
      required:
      - id
      - name
      - subsystems
      type: object
    SubsystemResponse:
      description: Данные подсистемы
      example:
        systemId: 7
        teamId: 9
        name: name
        description: description
        id: 2
      properties:
        id:
          description: Уникальный идентификатор подсистемы
          format: int64
          type: integer
        name:
          description: Название подсистемы
          type: string
        description:
          description: Описание подсистемы
          nullable: true
          type: string
        systemId:
          description: ID родительской системы
          format: int64
          type: integer
        teamId:
          description: ID ответственной команды
          format: int64
          type: integer
      required:
      - id
      - name
      - systemId
      - teamId
      type: object
    SystemRequest:
      description: Данные для создания/обновления системы
      example:
        name: name
        description: description
      properties:
        name:
          description: Название системы
          maxLength: 255
          minLength: 1
          type: string
        description:
          description: Описание системы
          maxLength: 1000
          nullable: true
          type: string
      required:
      - name
      type: object
    SubsystemPageResponse:
      description: Список подсистем
      items:
        $ref: "#/components/schemas/SubsystemResponse"
      type: array
    SubsystemRequest:
      description: Данные для создания/обновления подсистемы
      example:
        systemId: 0
        teamId: 6
        name: name
        description: description
      properties:
        name:
          description: Название подсистемы
          maxLength: 255
          minLength: 1
          type: string
        description:
          description: Описание подсистемы
          maxLength: 1000
          nullable: true
          type: string
        systemId:
          description: ID родительской системы
          format: int64
          type: integer
        teamId:
          description: ID ответственной команды
          format: int64
          type: integer
      required:
      - name
      - systemId
      - teamId
      type: object
...
    PageResponse:
      description: Общая схема для пагинированных ответов
      properties:
        totalElements:
          description: Общее количество элементов
          format: int64
          type: integer
        totalPages:
          description: Общее количество страниц
          type: integer
        size:
          description: Размер страницы
          type: integer
        number:
          description: Номер текущей страницы (начиная с 0)
          type: integer
        first:
          description: Является ли текущая страница первой
          type: boolean
        last:
          description: Является ли текущая страница последней
          type: boolean
      required:
      - first
      - last
      - number
      - size
      - totalElements
      - totalPages
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      description: JWT token
      scheme: bearer
      type: http